+++ src/main/java/com/buuz135/simpleclaims/Main.java
@@ -51,6 +51,7 @@ public class Main extends JavaPlugin {
         this.getEntityStoreRegistry().registerSystem(new PlaceBlockEventSystem());
         this.getEntityStoreRegistry().registerSystem(new InteractEventSystem());
         this.getEntityStoreRegistry().registerSystem(new PickupInteractEventSystem());
+        this.getEntityStoreRegistry().registerSystem(new TamedEntityDamageEventSystem());
         this.getEntityStoreRegistry().registerSystem(new TitleTickingSystem(CONFIG.get().getTitleTopClaimTitleText(), CONFIG.get().getWildernessName()));
         if (CONFIG.get().isEnableAlloyEntryTesting())
             this.getEntityStoreRegistry().registerSystem(new EntryTickingSystem());