+++ src/main/java/com/buuz135/simpleclaims/config/SimpleClaimsConfig.java
@@ -26,6 +26,7 @@ public class SimpleClaimsConfig {
     private boolean DefaultPartyInteractBench = false;
     private boolean DefaultPartyInteractChair = false;
     private boolean DefaultPartyInteractPortal = false;
+    private boolean DefaultPartyTamedDamageEnabled = false;
 
     private boolean AllowPartyPVPSetting = true;
     private boolean AllowPartyFriendlyFireSetting = true;
@@ -38,6 +39,7 @@ public class SimpleClaimsConfig {
     private boolean AllowPartyInteractBenchSetting = true;
     private boolean AllowPartyInteractChairSetting = true;
     private boolean AllowPartyInteractPortalSetting = true;
+    private boolean AllowPartyTamedDamageSetting = true;
 
     private String[] ClaimCommandAliases = new String[]{"sc", "sc-chunks", "scc", "chunk", "chunks", "claim", "claims"};
     private String[] WorldNameBlacklistForClaiming = new String[0];
@@ -59,6 +61,7 @@ public class SimpleClaimsConfig {
     private boolean EnablePerimeterReservation = false;
     private boolean ShowPerimeterReservationOnTheMap = false;
 
+    private String[] TamedEntityRoleIdentifiers = new String[]{"Tamed_"};
     private String[] BlocksThatIgnoreInteractRestrictions = new String[]{"gravestone"};
 
     public SimpleClaimsConfig() {
@@ -197,8 +200,16 @@ public class SimpleClaimsConfig {
         return DefaultPartyInteractPortal;
     }
 
+    public boolean isDefaultPartyTamedDamageEnabled() {
+        return DefaultPartyTamedDamageEnabled;
+    }
+
     public boolean isAllowPartyInteractPortalSetting() {
         return AllowPartyInteractPortalSetting;
+    }
+
+    public boolean isAllowPartyTamedDamageSetting() {
+        return AllowPartyTamedDamageSetting;
     }
 
     public boolean isEnableParticleBorders() {
@@ -207,6 +218,10 @@ public class SimpleClaimsConfig {
 
     public String[] getBlocksThatIgnoreInteractRestrictions() {
         return BlocksThatIgnoreInteractRestrictions;
+    }
+
+    public String[] getTamedEntityRoleIdentifiers() {
+        return TamedEntityRoleIdentifiers;
     }
 
     public int getMaxPartyMembers() {
@@ -260,4 +275,5 @@ public class SimpleClaimsConfig {
     public boolean isShowPerimeterReservationOnTheMap() {
         return ShowPerimeterReservationOnTheMap;
     }
+
 }