+++ src/main/java/com/buuz135/simpleclaims/gui/subscreens/InteractGui.java
@@ -66,6 +66,9 @@ public class InteractGui extends GuiWithParent<InteractGui.ChunkListGuiData> {
             if (data.action.equals("InteractPortalSetting")) {
                 this.partyInfo.setOverride(new PartyOverride(PartyOverrides.PARTY_PROTECTION_INTERACT_PORTAL, new PartyOverride.PartyOverrideValue("bool", !this.partyInfo.isPortalInteractEnabled())));
             }
+            if (data.action.equals("TamedDamageSetting")) {
+                this.partyInfo.setOverride(new PartyOverride(PartyOverrides.PARTY_PROTECTION_TAMED_DAMAGE, new PartyOverride.PartyOverrideValue("bool", !this.partyInfo.isTamedDamageEnabled())));
+            }
             ClaimManager.getInstance().saveParty(this.partyInfo);
             UICommandBuilder commandBuilder = new UICommandBuilder();
             UIEventBuilder eventBuilder = new UIEventBuilder();
@@ -103,6 +106,10 @@ public class InteractGui extends GuiWithParent<InteractGui.ChunkListGuiData> {
         uiCommandBuilder.set("#InteractPortalSetting #CheckBox.Value", this.partyInfo.isPortalInteractEnabled());
         uiCommandBuilder.set("#InteractPortalSetting #CheckBox.Disabled", !playerCanModify || !Main.CONFIG.get().isAllowPartyInteractPortalSetting());
         uiEventBuilder.addEventBinding(CustomUIEventBindingType.ValueChanged, "#InteractPortalSetting #CheckBox", EventData.of("Action", "InteractPortalSetting"), false);
+
+        uiCommandBuilder.set("#TamedDamageSetting #CheckBox.Value", this.partyInfo.isTamedDamageEnabled());
+        uiCommandBuilder.set("#TamedDamageSetting #CheckBox.Disabled", !playerCanModify || !Main.CONFIG.get().isAllowPartyTamedDamageSetting());
+        uiEventBuilder.addEventBinding(CustomUIEventBindingType.ValueChanged, "#TamedDamageSetting #CheckBox", EventData.of("Action", "TamedDamageSetting"), false);
     }
 
     public static class ChunkListGuiData extends GuiWithParentData {